export types (#204)
